+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -24,6 +24,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added dither transition between cascades for Low and Medium quality settings
 - Added single-pass instancing support with XR SDK
 - Added occlusion mesh support with XR SDK
+- Added support of Alembic velocity to various shaders
+- Added support for more than 2 views for single-pass instancing
 
 ### Fixed
 - Fixed an issue with history buffers causing effects like TAA or auto exposure to flicker when more than one camera was visible in the editor
@@ -57,6 +59,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed post-processing effect relying on Unity's random number generator
 - Fixed camera flickering when using TAA and selecting the camera in the editor
 - Fixed issue with single shadow debug view and volumetrics
+- Fixed indirect deferred compute with XR single-pass instancing
 
 ### Changed
 - Optimization: Reduce the group size of the deferred lighting pass from 16x16 to 8x8